add a new search parameter 'journal' for Inspire.
Most articles' references do not contain the =title= information.
Therefore only the =author= parameter could be used in bib-search and
obviously not enough. However, the =journal= parameters are always
listed in the reference page and as far as I know, many researchers name
their downloaded articles with the journal info.

For my usage, InspireHEP is called as the backend mostly and the
bib-info of that is the best I've known. Thus the =journal= parameter is
only added for this backend. arXiv maybe added later.